Updated main.cpp
Added exceptions in "binaryToDecimal" function for input that is not
binary, and for input that is larger than 8 digits.

Added exceptins in "hexToDecimal" for inputs that is not a hexadecimal
value (0-9, a-f), and for input that is larger than 4 digits.

Updated the switch statement cases for both functions to include
try/catch blocks that handle the given exceptions.