Updated main.cpp
Added an enumerated data type, along with an std::map, which allows the
switch statement to take in string inputs!

This prevents certain errors that can arise when reading in an integer
or char input for the switch's menu choice. For example, if a user was
to type "1xyzabc", previously this would go into case 1 of the switch
statement, but then give an error due to the invalid input. Now, this is
caught by the map and enum combination, and gives the correct error in
the "default" case of the switch statement.